add feature: send_message_as_string
