+++ src/controller/get_proposal_data_controller.rs
@@ -1,6 +1,6 @@
 use crate::presentation::pretty_json_presenter::print_pretty_json;
-use crate::presentation::send_message::send_message;
-use crate::presentation::send_message::RunningStatus;
+use crate::presentation::send_message_to_console::send_message_to_console;
+use crate::presentation::send_message_to_console::RunningStatus;
 use crate::use_case::build_structured_information::builder::build_structured_proposal_information;
 use crate::use_case::download_html_page::downloader::download_html_page;
 use crate::use_case::download_og_image::downloader::download_og_image;
@@ -10,7 +10,7 @@ pub fn get_proposal_data_controller(url: &str) {
      * Download HTML page from the given URL
      */
     download_html_page(url);
-    send_message(
+    send_message_to_console(
         RunningStatus::Success,
         "HTML page has been successfully downloaded.",
     );
@@ -19,7 +19,7 @@ pub fn get_proposal_data_controller(url: &str) {
      * Extract structured information from the downloaded HTML page
      */
     let proposal = build_structured_proposal_information();
-    send_message(
+    send_message_to_console(
         RunningStatus::Success,
         "Structured proposal information has been successfully created.",
     );
@@ -40,7 +40,7 @@ pub fn get_proposal_data_controller(url: &str) {
             return;
         }
     };
-    send_message(
+    send_message_to_console(
         RunningStatus::Success,
         "OG Image has been successfully downloaded.",
     );
@@ -48,7 +48,7 @@ pub fn get_proposal_data_controller(url: &str) {
     /*
      * Show how to get the downloaded image for users
      */
-    send_message(
+    send_message_to_console(
         RunningStatus::Notice,
         format!(
             "you can get data by running: cp `{}` path/your/directory",
