+++ src/presentation.rs
@@ -1,4 +1,4 @@
 pub mod pretty_json_presenter;
-pub mod send_message;
+pub mod send_message_to_console;
 pub mod show_json;
 pub mod terminal_message_presenter;
