+++ src/use_case/build_structured_information/builder.rs
@@ -9,10 +9,10 @@ use crate::infrastructure::file_path_provider::html_file_path_provider::HtmlFile
 use crate::infrastructure::file_path_provider::json_file_path_provider::JsonFilePathProvider;
 use crate::infrastructure::reader::read_html::read_html;
 use crate::infrastructure::writer::write_json_from_proposal::write_json_from_proposal;
+use crate::presentation::send_message_to_console::send_message_to_console;
+use crate::presentation::send_message_to_console::RunningStatus;
 use crate::presentation::terminal_message_presenter::ConsoleMessenger;
 use crate::presentation::terminal_message_presenter::MessageType;
-use crate::presentation::send_message::RunningStatus;
-use crate::presentation::send_message::send_message;
 
 pub fn build_structured_proposal_information() -> ProposalModel {
     let error_message = ConsoleMessenger::new(
@@ -26,11 +26,8 @@ pub fn build_structured_proposal_information() -> ProposalModel {
     let html_path_provider = HtmlFilePathProvider::new("proposal");
     let html_path = html_path_provider.get_path();
 
-    let html_text = read_html(html_path)
-        .unwrap_or_else(|_| panic!(
-            "{}",
-            error_message.supply_message()
-        ));
+    let html_text =
+        read_html(html_path).unwrap_or_else(|_| panic!("{}", error_message.supply_message()));
 
     /*
      * extract title and speaker from html
@@ -53,10 +50,10 @@ pub fn build_structured_proposal_information() -> ProposalModel {
         speaker,
         og_image_url,
     };
-    
-    send_message(
+
+    send_message_to_console(
         RunningStatus::Success,
-        "Successfully get structured information from the HTML file"
+        "Successfully get structured information from the HTML file",
     );
 
     /*
